refactor(server): split route_bridge delivery and registration into submodules
Code-motion-only follow-up to review: delivery/mod.rs (1,650 lines)
becomes mod/local/ordered/remote alongside the existing muc and
receiver modules; registration.rs (1,459 lines) becomes
registration/{mod,socket,owner,side_effects}. No file exceeds 778
lines; function inventory is unchanged and external paths are
preserved via re-exports.
